feat: 노드 삭제 시도시, 처리중인 이미지가 있을 경우 Graceful 하게 삭제되도록 대응

+++ eagle-film-consumer/src/main/kotlin/org/gdsc/yonsei/eagleflim/consumer/service/NodeService.kt
@@ -22,8 +22,11 @@ class NodeService(
 	}
 
 	fun removeNode(address: String) {
-		if (!nodeRepository.checkNode(address)) {
-			error("node not exist")
+		val nodeInfo = nodeRepository.getNodeInfo(address) ?: error("node not exist")
+
+		if (nodeInfo.assignedRequest != null) {
+			nodeRepository.insertDeleteQueue(address)
+			return
 		}
 
 		nodeRepository.removeNode(address)
